Organized the Files
Moved around files and deleted a duplicate.
