+++ Test/.gitignore
@@ -1,0 +1,4 @@
+/JsonUtilsTest.class
+/UmlClassTest.class
+/UmlEditorTest.class
+/UmlRelationshipTest.class
